Change floormanual function to accept long type

+++ soal1.c
@@ -12,7 +12,7 @@
  #include <string.h>
 
 
- long floormanual(double x) {
+ int floormanual(long x) {
     int i = (int)x; // i = integralnya saja dari x
     if (x < 0 && x != i) {
         return i - 1; // situasi negatif
